feat: layoutファイルのみ利用出来るようになりました。
